Added Nail Button
New object can be placed in rooms. If a block is place on the button it can activate a connect object. Currently the only connect object is a moving platform, I think I need to make a public class to parent the moving platform to? Also you have to connect the ground block part of the to the button.