fixed wall boxes
I changed the grid composite collider from outlines to polygons so the blocks arn't as hollow and it seemed to fix being able to place boxes inside walls. This could be better fixed by implementing a raycast feature, which would also stop placing boxes on the other side of walls.